# src/main/java/com/example/absservice/swagger/api/ClientsApi.java
/**
 * NOTE: This class is auto generated by the swagger code generator program (3.0.56).
 * https://github.com/swagger-api/swagger-codegen
 * Do not edit the class manually.
 */

package com.example.absservice.swagger.api;

import java.io.*;
import java.util.*;

import javax.validation.*;
import javax.validation.constraints.*;

import jakarta.servlet.http.HttpServletRequest;
import org.slf4j.*;

import com.fasterxml.jackson.databind.*;
import io.swagger.v3.oas.annotations.*;
import io.swagger.v3.oas.annotations.enums.*;
import io.swagger.v3.oas.annotations.media.*;
import io.swagger.v3.oas.annotations.responses.*;
import io.swagger.v3.oas.annotations.security.*;
import org.springframework.http.*;
import org.springframework.validation.annotation.*;
import org.springframework.web.bind.annotation.*;

import com.example.absservice.swagger.dto.*;

@Validated
public interface ClientsApi {

    Logger log = LoggerFactory.getLogger(ClientsApi.class);

    default Optional<ObjectMapper> getObjectMapper() {
        return Optional.empty();
    }

    default Optional<HttpServletRequest> getRequest() {
        return Optional.empty();
    }

    default Optional<String> getAcceptHeader() {
        return getRequest().map(r -> r.getHeader("Accept"));
    }

    @Operation(summary = "Метод идентификации клиента по пасспорту", description = "", security = {
        @SecurityRequirement(name = "BearerAuth")}, tags = {"ABS"})
    @ApiResponses(value = {
        @ApiResponse(responseCode = "200", description = "Успешный ответ с найденным клиентом", content = @Content(mediaType = "application/json", schema = @Schema(implementation = ClientDto.class))),

        @ApiResponse(responseCode = "400", description = "Bad Request", content = @Content(mediaType = "application/json", schema = @Schema(implementation = ApiErrorResponseDto.class))),

        @ApiResponse(responseCode = "401", description = "Unauthorized", content = @Content(mediaType = "application/json", schema = @Schema(implementation = ApiErrorResponseDto.class))),

        @ApiResponse(responseCode = "404", description = "Not Found", content = @Content(mediaType = "application/json", schema = @Schema(implementation = ApiErrorResponseDto.class))),

        @ApiResponse(responseCode = "500", description = "Internal Server Error", content = @Content(mediaType = "application/json", schema = @Schema(implementation = ApiErrorResponseDto.class)))})
    @RequestMapping(value = "/clients",
        produces = {"application/json"},
        method = RequestMethod.GET)
    default ResponseEntity<ClientDto> getClient(
        @NotNull @Pattern(regexp = "^[0-9]+$") @Size(min = 10, max = 10) @Parameter(in = ParameterIn.QUERY, description = "Данные паспорта (серия номер) для идентификации клиента", required = true, schema = @Schema()) @Valid @RequestParam(value = "passport", required = true) String passport
    ) {
        if (getObjectMapper().isPresent() && getAcceptHeader().isPresent()) {
            if (getAcceptHeader().get().contains("application/json")) {
                try {
                    return new ResponseEntity<>(getObjectMapper().get()
                        .readValue(
                            "{\r\n  \"passport\" : \"0000000000\",\r\n  \"name\" : \"Иванов Иван Иванович\",\r\n  \"categories\" : [ \"Сотрудник\", \"Клиент\" ],\r\n  \"birthDate\" : \"1985-05-15T00:00:00.000+00:00\"\r\n}",
                            ClientDto.class), HttpStatus.NOT_IMPLEMENTED);
                } catch (IOException e) {
                    log.error("Couldn't serialize response for content type application/json", e);
                    return new ResponseEntity<>(HttpStatus.INTERNAL_SERVER_ERROR);
                }
            }
        } else {
            log.warn(
                "ObjectMapper or HttpServletRequest not configured in default ClientsApi interface so no example is generated");
        }
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);
    }

}