# src/main/resources/swagger-api/abs_swagger.yaml
openapi: 3.0.0
info:
  title: Калькулятор депозитных продуктов
  version: 2.0.1
servers:
  - url: http://localhost:18082/api/v2/abs
...
paths:
  /clients:
    get:
      security:
        - BearerAuth: []
      summary: Метод идентификации клиента по пасспорту
      tags:
        - ABS
      operationId: getClient
      parameters:
        - in: query
          name: passport
          description: Данные паспорта (серия номер) для идентификации клиента
          schema:
            type: string
            minLength: 10
            maxLength: 10
            pattern: ^[0-9]+$
          example: '0000000000'
          required: true
      responses:
        '200':
          description: Успешный ответ с найденным клиентом
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ClientDto'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponseDto'
              examples:
                example1:
                  summary: Пример ошибки с кодом 2001
                  value:
                    code: 2001
                    message: 'Ошибка запроса,  отсутствуют следующие параметры запроса: passport.'
                example2:
                  summary: Пример ошибки с кодом 2002
                  value:
                    code: 2001
                    message: 'Ошибка запроса,  некорректные данные в следующих параметрах: passport - 0000132.'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponseDto'
              example:
                code: 2003
                message: 'AccessToken не валиден'

        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponseDto'
              example:
                code: 2004
                message: 'Данные о клиенте с указанным идентификатором не найдены'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponseDto'
              example:
                code: 500
                message: Внутреняя ошибка сервера или базы данных
...
components:
  schemas:
...
    ClientDto:
      type: object
      properties:
        categories:
          type: array
          items:
            type: string
          example:
            - Сотрудник
            - Клиент
        name:
          type: string
          description: Фамилия Имя Отчество
          example: Иванов Иван Иванович
        birthDate:
          type: string
          description: Дата рождения клиента, формат ГГГГ-ММ-ДД
          format: date
          example: '1985-05-15'
        passport:
          type: string
          description: Серия и номер паспорта клиента, 10 цифр
          example: '0000000000'
...
    ApiErrorResponseDto:
      type: object
      properties:
        code:
          type: integer
        message:
          type: string
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer